Display server and GUI terminal application
Commit also includes the 'poll' syscall, a new kernel idle task, and
several bug fixes related to VFS. A new event system was created for
hardware devices such as the mouse and keyboard.

The display server is currently very primitive and has some major bugs:
Mouse flickering, random crashes when resizing canvases for windows, and
some issues related to the order of which events are sent to
applications.